+++ src/parse.h
@@ -10,7 +10,7 @@ char *ltrim(char *s);
 char *rtrim(char *s);
 
 char *parse_str(FILE *stream);
-bool parse_llint(FILE *stream, long long int *r);
+bool parse_llint(FILE *stream, long int *r);
 void skip_white(FILE *stream);
 
 #endif
